% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-polygon-area.R
\name{compute_polygon_area}
\alias{compute_polygon_area}
\title{Compute polygon area}
\usage{
compute_polygon_area(vertex.df, signed = FALSE)
}
\arguments{
\item{vertex.df}{A data frame where each row corresponds to the vertex of a polygon.
It must contain the columns \code{x}, \code{y}, and optionally \code{group} (must be integer).
\code{x} and \code{y} specify the coordinates of the vertex,
and \code{group} is used to indicate which polygon the vertex belongs to.
If \code{vertex.df} contains only a single polygon, \code{group} can be omitted.}

\item{signed}{If true, the signed areas of the polygons are returned (positive for
counter-clockwise oriented polygons, negative for clockwise oriented.)}
}
\value{
A data frame that contains two columns:
\item{group}{The group number of the polygon}
\item{area}{The area of the polygon}
}
\description{
This function takes in a data frame containing one or more polygons,
and computes the (signed) area of each polygon.
}
\examples{
# Example with multiple polygons
vertex.df = compute_regular_polygons(
  center = data.frame(x = c(0, 10), y = c(0, 0)),
  radius = 1,
  num.edges = c(4, 1000)
)
area.df = compute_polygon_area(vertex.df)

# Note that the area of polygon 2 is close to pi!
print(area.df)

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
