% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-polygon-perimeter.R
\name{compute_polygon_perimeter}
\alias{compute_polygon_perimeter}
\title{Compute polygon perimeter}
\usage{
compute_polygon_perimeter(vertex.df)
}
\arguments{
\item{vertex.df}{A data frame where each row corresponds to the vertex of a polygon.
It must contain the columns \code{x}, \code{y}, and optionally \code{group} (must be integer).
\code{x} and \code{y} specify the coordinates of the vertex,
and \code{group} is used to indicate which polygon the vertex belongs to.
If \code{vertex.df} contains only a single polygon, \code{group} can be omitted.}
}
\value{
A data frame that contains two columns:
\item{group}{The group number of the polygon}
\item{perimeter}{The perimeter of the polygon}
}
\description{
This function takes in a data frame containing one or more polygons,
and computes the perimeter of each polygon
}
\examples{
# Example with multiple polygons
vertex.df = compute_regular_polygons(
  center = data.frame(x = c(0, 10), y = c(0, 0)),
  radius = 1,
  num.edges = c(4, 1000)
)
perimeter.df = compute_polygon_perimeter(vertex.df)

# The perimeters should be around 4*sqrt(2) and 2*pi
print(perimeter.df)

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
