% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-regular-polygons.R
\name{compute_regular_polygons}
\alias{compute_regular_polygons}
\title{Compute regular polygons}
\usage{
compute_regular_polygons(
  center = c(0, 0),
  radius = 1,
  rotation = 0,
  num.edges = 50,
  include.info = TRUE,
  ...
)
}
\arguments{
\item{center}{Specifies the center(s) of the polygon(s). Can either be a
numeric vector of length 2 (every polygon has the same center), or a
data frame/matrix with 1 row per polygon. If the data frame/matrix has
columns named \code{x} and \code{y}, these are used as center coordinates.
Otherwise, the first two columns are used.}

\item{radius}{The radius of the circumscribed circle of the polygon. Can either
be a single number or a vector containing one radius per polygon.}

\item{rotation}{The amount of rotation applied to the polygon(s), in radians.
If \code{rotation = 0}, then the first vertex is at the "top" of the polygon.
Can either be a single number or a vector containing a different rotation per polygon.}

\item{num.edges}{The number of edges that the polygon(s) contain. Can either
be a single integer or a vector containing a different number of edges per polygon.}

\item{include.info}{If true, the returned data frame will contain the values for
\code{radius}, \code{rotation} and \code{num.edges} as columns.}

\item{...}{Custom properties that can be used to store additional information about
the polygon(s), such as color and opacity. Can either be a single value or a
vector containing a different value per polygon.}
}
\value{
A data frame where each row corresponds to the vertex of a polygon.
The following columns are included:
\item{x, y}{The coordinates of the vertex}
\item{group}{The group number of the polygon}
\item{radius, rotation, num.edges}{ The parameters that are specified in the function call.
Only included if \code{include.info = TRUE}}
\item{...}{Custom columns that are specified in the function call.}
}
\description{
This function computes one or more regular polygons.
}
\examples{
set.seed(123)
# This generates 10 polygons with different centers, an individual number of edges
# and a custom property "color"
vertex.df = compute_regular_polygons(
  center = data.frame(x = 1:10, y = 1:10),
  radius = 0.5,
  num.edges = 3:12,
  color = sample(c("red", "blue", "green"), 10, replace = TRUE)
)

library(ggplot2)
ggplot()+
  geom_polygon(data = vertex.df, aes(x = x, y = y, group = group, fill = color))+
  scale_fill_identity()+
  coord_fixed()


}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
