% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-polygons.R
\name{cut_polygons}
\alias{cut_polygons}
\title{Repeated cutting of polygons}
\usage{
cut_polygons(
  vertex.df,
  number.of.iterations,
  use.centroid = TRUE,
  seed = NULL,
  return.all = FALSE
)
}
\arguments{
\item{vertex.df}{A data frame where each row corresponds to the vertex of a polygon.
It must contain the columns \code{x}, \code{y}, and optionally \code{group} (must be integer).
\code{x} and \code{y} specify the coordinates of the vertex,
and \code{group} is used to indicate which polygon the vertex belongs to.
If \code{vertex.df} contains only a single polygon, \code{group} can be omitted.}

\item{number.of.iterations}{The number of times that the polygons are cut in
half. Either a single value, or a numeric vector with one value per polygon in \code{vertex.df}.}

\item{use.centroid}{If \code{TRUE}, the randomly generated line goes through the centroid
of the polygon. Otherwise, the line goes through a point sampled uniformly
from the interior of the polygon.}

\item{seed}{The seed used when generating the directions of the cut lines.
When no value is specified, the seed is not set.}

\item{return.all}{If \code{TRUE}, the function will return a data frame
that contains every iteration (i.e. the state after iteration 1,...,\code{number.of.iterations}).}
}
\value{
A data frame where each row corresponds to the vertex of a polygon,
with the following columns:
\item{x, y}{The coordinates of the vertex}
\item{group}{The index of the polygon that the vertex belongs to}
\item{iteration}{The iteration that the polygon belongs to (only present when \code{return.all = TRUE})}
}
\description{
This function takes in an arbitrary number of polygons and cuts them a given
number of times. For each iteration, every polygon is cut by a randomly generated
line, which divides it into several parts.
}
\details{
The directions of the lines are sampled from a uniform distribution.
}
\note{
\code{number.of.iterations} should be chosen with care, as each
iteration at least doubles the number of polygons, leading to exponential growth.
}
\examples{
# Example with single initial polygon (square)
vertex.df = data.frame(
  x = c(-1, 1, 1, -1),
  y = c(1, 1, -1, -1)
)
cut.df = cut_polygons(vertex.df, number.of.iterations = 10, seed = 123)

library(ggplot2)
ggplot()+
  geom_polygon(data = cut.df,
               aes(x = x, y = y, group = group),
               fill = NA, color = "black")+
  coord_fixed()

# Example with multiple initial polygons (square and triangle)
vertex.df = data.frame(
  x = c(-1, 1, 1, -1, 1, -1, 0),
  y = c(1, 1, -1, -1, 2, 2, 2+sqrt(3)),
  group = c(rep(1, 4), rep(2, 3))
)

# Cut square and triangle 10 and 9 times, respectively
cut.df = cut_polygons(vertex.df, number.of.iterations = c(10, 9), seed = 123)

ggplot()+
  geom_polygon(data = cut.df,
               aes(x = x, y = y, group = group),
               fill = NA, color = "black")+
  coord_fixed()

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
