% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_polygon.R
\name{generate_random_polygon}
\alias{generate_random_polygon}
\title{Generate random polygon}
\usage{
generate_random_polygon(
  center = c(0, 0),
  min.radius = 0.5,
  max.radius = 1,
  num.vertices = 5,
  seed = NULL
)
}
\arguments{
\item{center}{The "center" of the polygon, which determines its position. Should
either be a 1 x 2 data frame/matrix or a numeric vector with 2 elements.}

\item{min.radius, max.radius}{The minimum and maximum bounds used in the uniform
distribution that samples the radius values}

\item{num.vertices}{The number of vertices that the random polygon should contain}

\item{seed}{The seed used when sampling the radius value}
}
\value{
A data frame that contains one row per vertex in the polygon, and the
columns \code{x} and \code{y}.
}
\description{
This function generates a random polygon. It uses the polar coordinate
representation, where the angles are evenly spaced over the interval [0, 2pi]
and the radiuses are sampled from a uniform distribution.
}
\examples{
# A random polygon with center (5, 10) and 10 vertices. The distance from
# the center to the vertices is between 4 and 10
vertex.df = generate_random_polygon(
  center = c(5, 10),
  min.radius = 4,
  max.radius = 10,
  num.vertices = 10,
  seed = 123
)

library(ggplot2)
# The center of the polygon is shown in red
ggplot()+
  geom_polygon(data = vertex.df, aes(x = x, y = y))+
  geom_point(aes(x = 5, y = 10), color = "red")+
  coord_fixed()

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
