% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax-scaling.R
\name{minmax_scaling}
\alias{minmax_scaling}
\title{Rescale values using min-max scaling}
\usage{
minmax_scaling(values, lower = 0, upper = 1)
}
\arguments{
\item{values}{A numeric vector of values}

\item{lower, upper}{The lower and upper bounds of the range that the values
are mapped onto}
}
\value{
A numeric vector that contains the rescaled values
}
\description{
Applies a rescaling that linearly transforms the values into a specified
range of values.
}
\examples{
set.seed(123)
# Generate values from normal distribution
original.values = rnorm(100, -4, 5)
# The values lie between approximately -15.5 and 7
hist(original.values)
# Rescale values to the interval [0, 1]
rescaled.values = minmax_scaling(original.values)
hist(rescaled.values)

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
