% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-vectors.R
\name{normalize_vectors}
\alias{normalize_vectors}
\title{Normalize vectors to have length 1}
\usage{
normalize_vectors(vectors)
}
\arguments{
\item{vectors}{A n x d matrix or data frame that contains one vector per row,
and one column per dimension.}
}
\value{
The return value is of the same type as \code{vectors}, and contains
the normalized vectors.
}
\description{
Takes in one or more vectors and normalizes their lengths to 1.
}
\examples{
set.seed(123)
# Generate random 2D vectors
vectors = matrix(runif(200, -1, 1), ncol = 2)
# The lengths are varied
summary(sqrt(rowSums(vectors^2)))
# Normalize vectors
normalized.vectors = normalize_vectors(vectors)
# The lengths are all equal to 1
summary(sqrt(rowSums(normalized.vectors^2)))

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
