% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounded-rectangle.R
\name{rounded_rectangle}
\alias{rounded_rectangle}
\title{Generate rectangle with rounded corners}
\usage{
rounded_rectangle(
  center = c(0, 0),
  width = 1,
  height = 1,
  rotation = 0,
  corner.radius = NULL,
  vertices.per.corner = 50
)
}
\arguments{
\item{center}{The center/centroid of the shape. Can either be a 1 x 2 data
frame/matrix or a numeric vector length 2.}

\item{width, height}{The width and height of the rectangle before the corners
are rounded.}

\item{rotation}{The amount of rotation applied to the shape, in radians.}

\item{corner.radius}{The radius of the circular arcs. Should be less than min(width, height) / 2.}

\item{vertices.per.corner}{The number of vertices that each circular arc consists of.}
}
\value{
A data frame containing the columns \code{x} and \code{y}, where each
row is a vertex in the rounded rectangle.
}
\description{
This function computes the vertices of a rectangle with rounded corners.
}
\note{
The result may contain less than \code{4*vertices.per.corner} vertices.
In certain situations, consecutive vertices will end up having the same coordinates.
The duplicates are then removed.
}
\examples{
# Create a rounded rectangle
vertex.df = rounded_rectangle(
  center = c(2, 5),
  width = 3,
  height = 6,
  rotation = pi/4,
  corner.radius = 1
)

# How the rounded rectangle looks:
library(ggplot2)
ggplot()+
  geom_polygon(data = vertex.df, aes(x = x, y = y))+
  coord_fixed()

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
