% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounded-regular-polygon.R
\name{rounded_regular_polygon}
\alias{rounded_regular_polygon}
\title{Generate regular polygon with rounded corners}
\usage{
rounded_regular_polygon(
  num.edges = 3,
  center = c(0, 0),
  radius = 1,
  rotation = 0,
  corner.radius = 0.25,
  vertices.per.corner = 50
)
}
\arguments{
\item{num.edges}{The number of edges that the regular polygon contains.}

\item{center}{The center/centroid of the shape. Can either be a 1 x 2 data
frame/matrix or a numeric vector length 2.}

\item{radius}{The radius of the circumscribed circle of the regular polygon.}

\item{rotation}{The amount of rotation applied to the shape, in radians.}

\item{corner.radius}{The radius of the circular arcs.}

\item{vertices.per.corner}{The number of vertices that each circular arc consists of.}
}
\value{
A data frame containing the columns \code{x} and \code{y}, where each
row is a vertex in the rounded polygon.
}
\description{
This function computes the vertices of a regular polygon with rounded corners.
}
\note{
The result may contain less than \code{num.edges*vertices.per.corner} vertices.
In certain situations, consecutive vertices will end up having the same coordinates.
The duplicates are then removed.
}
\examples{
# Create a rounded hexagon
vertex.df = rounded_regular_polygon(
  num.edges = 6,
  center = c(5, 6),
  radius = 2,
  rotation = pi/4,
  corner.radius = 0.5
)

# How the rounded hexagon looks:
library(ggplot2)
ggplot()+
  geom_polygon(data = vertex.df, aes(x = x, y = y))+
  coord_fixed()

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
