% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform-scaling.R
\name{uniform_scaling}
\alias{uniform_scaling}
\title{Transform values to be uniformly distributed}
\usage{
uniform_scaling(values)
}
\arguments{
\item{values}{A numeric vector of values}
}
\value{
A numeric vector that contains the rescaled values
}
\description{
Applies a rescaling that transforms the values into being uniformly
distributed on the interval [0, 1].
}
\examples{
set.seed(123)
# Generate values from normal distribution
original.values = rnorm(100, -4, 5)
# The values lie between approximately -15.5 and 7
hist(original.values)
# Rescale values uniformly to the interval [0, 1]
uniform.values = uniform_scaling(original.values)
hist(uniform.values)

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
